droiddec: Workaround some odd video width.
Some videos have a width which needs to be padded to allow gecko to render it properly
An example of such videos is http://www.youtube.com/watch?v=nG6NLvoltws

The video has a width of 406 pixels but a proper I420 video should have been 408 wide.
Our color converter does not have issues with converting such video. It however does
not add the needed padding. Gecko however does expect such padding to be there so we
just add it to please gecko.

It's a slower path which involves a memcpy() but we don't have another option.

Another example is http://www.w3.org/2010/05/video/mediaevents.html whose width
should be padded from 854 to 856