droidcamsrc: rework how we start video recording a bit to accommodate camerabin
We need to push FLUSH_STOP before we return from start_capture because camerabin will
push any application tags immediately which might fail if we are not fast enough.
If we don't do that then old tags might be used for new videos
We need also to set ready-for-capture to false to allow camerabin to set filesink to PLAYING
which will allow our FLUSH_STOP to propagate.